add unit test/change package name
