--- autopayroll-CordaService/contracts/src/main/kotlin/com/autopayroll/states/PaymentRequestState.kt
@@ -1,16 +1,0 @@
-package com.autopayroll.states
-
-import com.autopayroll.contracts.PaymentRequestContract
-import net.corda.core.contracts.BelongsToContract
-import net.corda.core.contracts.ContractState
-import net.corda.core.identity.AbstractParty
-import net.corda.core.identity.Party
-
-// *********
-// * State *
-// *********
-@BelongsToContract(PaymentRequestContract::class)
-data class PaymentRequestState(
-        val amount: String,
-        val toWhom: Party,
-        override val participants: List<AbstractParty> = listOf()) : ContractState
