--- autopayroll-CordaService/contracts/src/test/kotlin/com/autopayroll/contracts/ContractTests.kt
@@ -1,9 +1,0 @@
-package com.autopayroll.contracts
-
-import net.corda.testing.node.MockServices
-import org.junit.Test
-
-class ContractTests {
-    private val ledgerServices = MockServices()
-
-}
