+++ autopayroll-CordaService/gradle.properties
@@ -1,4 +1,4 @@
 name=Test
-group=com.autopayroll
+group=net.corda.examples.autopayroll
 version=0.1
 kotlin.incremental=false
