--- autopayroll-CordaService/workflows/src/test/kotlin/com/autopayroll/ContractTests.kt
@@ -1,9 +1,0 @@
-package com.autopayroll
-
-import net.corda.testing.node.MockServices
-import org.junit.Test
-
-class ContractTests {
-    private val ledgerServices = MockServices()
-
-}
