--- autopayroll-CordaService/workflows/src/test/kotlin/com/autopayroll/FlowTests.kt
@@ -1,35 +1,0 @@
-package com.autopayroll
-
-import com.autopayroll.flows.RequestFlowResponder
-import net.corda.testing.node.MockNetwork
-import net.corda.testing.node.MockNetworkParameters
-import net.corda.testing.node.TestCordapp
-import org.junit.After
-import org.junit.Before
-import org.junit.Test
-
-class FlowTests {
-    private val network = MockNetwork(MockNetworkParameters(cordappsForAllNodes = listOf(
-        TestCordapp.findCordapp("com.autopayroll.contracts"),
-        TestCordapp.findCordapp("com.autopayroll.flows")
-    )))
-    private val a = network.createNode()
-    private val b = network.createNode()
-
-    init {
-        listOf(a, b).forEach {
-            it.registerInitiatedFlow(RequestFlowResponder::class.java)
-        }
-    }
-
-    @Before
-    fun setup() = network.runNetwork()
-
-    @After
-    fun tearDown() = network.stopNodes()
-
-//    @Test
-//    fun `dummy test`() {
-//
-//    }
-}
