--- autopayroll-H2Database/TRADEMARK
@@ -1,4 +1,0 @@
-Corda and the Corda logo are trademarks of R3CEV LLC and its affiliates. All rights reserved.
-
-For R3CEV LLC's trademark and logo usage information, please consult our Trademark Usage Policy at
-https://www.r3.com/trademark-policy/.
