--- autopayroll-H2Database/config/dev/log4j2.xml
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="info">
-
-    <Properties>
-        <Property name="log-path">logs</Property>
-        <Property name="log-name">node-${hostName}</Property>
-        <Property name="archive">${log-path}/archive</Property>
-    </Properties>
-
-    <ThresholdFilter level="trace"/>
-
-    <Appenders>
-        <Console name="Console-Appender" target="SYSTEM_OUT">
-            <PatternLayout>
-                <pattern>
-                    %highlight{%level{length=1} %d{HH:mm:ss} %T %c{1}.%M - %msg%n}{INFO=white,WARN=red,FATAL=bright red blink}
-                </pattern>>
-            </PatternLayout>
-        </Console>
-
-        <!-- Will generate up to 10 log files for a given day. During every rollover it will delete
-             those that are older than 60 days, but keep the most recent 10 GB -->
-        <RollingFile name="RollingFile-Appender"
-                     fileName="${log-path}/${log-name}.log"
-                     filePattern="${archive}/${log-name}.%d{yyyy-MM-dd}-%i.log.gz">
-
-            <PatternLayout pattern="[%-5level] %d{ISO8601}{GMT+0} [%t] %c{1} - %msg%n"/>
-
-            <Policies>
-                <TimeBasedTriggeringPolicy/>
-                <SizeBasedTriggeringPolicy size="10MB"/>
-            </Policies>
-
-            <DefaultRolloverStrategy min="1" max="10">
-                <Delete basePath="${archive}" maxDepth="1">
-                    <IfFileName glob="${log-name}*.log.gz"/>
-                    <IfLastModified age="60d">
-                        <IfAny>
-                            <IfAccumulatedFileSize exceeds="10 GB"/>
-                        </IfAny>
-                    </IfLastModified>
-                </Delete>
-            </DefaultRolloverStrategy>
-
-        </RollingFile>
-    </Appenders>
-
-    <Loggers>
-        <Root level="info">
-            <AppenderRef ref="Console-Appender"/>
-            <AppenderRef ref="RollingFile-Appender"/>
-        </Root>
-        <Logger name="net.corda" level="info" additivity="false">
-            <AppenderRef ref="Console-Appender"/>
-            <AppenderRef ref="RollingFile-Appender"/>
-        </Logger>
-    </Loggers>
-
-</Configuration>
