--- autopayroll-H2Database/config/test/log4j2.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="info">
-    <Appenders>
-        <Console name="Console-Appender" target="SYSTEM_OUT">
-            <PatternLayout>
-                <pattern>
-                    [%-5level] %d{HH:mm:ss.SSS} [%t] %c{1}.%M - %msg%n
-                </pattern>>
-            </PatternLayout>
-        </Console>
-    </Appenders>
-    <Loggers>
-        <Root level="info">
-            <AppenderRef ref="Console-Appender"/>
-        </Root>
-        <Logger name="net.corda" level="info" additivity="false">
-            <AppenderRef ref="Console-Appender"/>
-        </Logger>
-    </Loggers>
-</Configuration>
