--- autopayroll-H2Database/contracts/build.gradle
@@ -1,21 +1,0 @@
-apply plugin: 'net.corda.plugins.cordapp'
-
-cordapp {
-    targetPlatformVersion corda_platform_version
-    minimumPlatformVersion corda_platform_version
-    contract {
-        name "AutoPayroll H2Demo Contracts"
-        vendor "Corda Open Source"
-        licence "Apache License, Version 2.0"
-        versionId 1
-    }
-}
-
-dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-
-    // Corda dependencies.
-    cordaCompile "$corda_release_group:corda-core:$corda_release_version"
-
-    testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
-}
