--- autopayroll-H2Database/gradle.properties
@@ -1,4 +1,0 @@
-name=Test
-group=com.autopayroll
-version=0.1
-kotlin.incremental=false
