--- autopayroll-H2Database/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Fri Aug 25 12:50:39 BST 2017
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.4.1-all.zip
