--- autopayroll-H2Database/settings.gradle
@@ -1,2 +1,0 @@
-include 'workflows'
-include 'contracts'
