--- autopayroll-H2Database/workflows/src/test/kotlin/com/autopayroll/ContractTests.kt
@@ -1,13 +1,0 @@
-package com.autopayroll
-
-import net.corda.testing.node.MockServices
-import org.junit.Test
-
-class ContractTests {
-    private val ledgerServices = MockServices()
-
-//    @Test
-//    fun `dummy test`() {
-//
-//    }
-}
